extracted more logic from Practice4 component to custom hook
